+++ src/main/java/com/plane/article/dto/ArticleInteractionRequset.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Pattern;
 public class ArticleInteractionRequset {
 
 	@NotNull(message = "글 번호는 필수입니다.")
-	private Integer articleId;
+	private Long articleId;
 	
 	
 	@Pattern(regexp = "^(SAVE|RECOMMAND)$", message = "상호작용은 SAVE, RECOMMAND 중 하나여야 합니다")
@@ -17,12 +17,12 @@ public class ArticleInteractionRequset {
 	}
 
 
-	public Integer getArticleId() {
+	public Long getArticleId() {
 		return articleId;
 	}
 
 
-	public void setArticleId(Integer articleId) {
+	public void setArticleId(Long articleId) {
 		this.articleId = articleId;
 	}
